Mathematical modelling: Forecasting cancer
Cancer's heterogeneity is the main reason it is so difficult to treat — if it retains a diverse enough population, a few impervious tumour cells can survive chemotherapy or radiation and re-establish their hold on the body. So mathematicians are creating models to predict in which direction a mutating cancer cell will change, in the hope of stopping evolution in its tracks.